particle-engine: Introduce a 'particles per ms' rate limit
This attempts to address the problem of timestamp granularity raised in
b7eab5f by using a 'new particles per ms' rate limiter. With each update
of the particle engine, the time difference between the current time and
last update time is multiplied by this constant to provide a maximum
number of particles which should be created in this update. Once this
maximum has been reached, no more particles are created in this update.