particle-engine: Add a priv struct
The idea here is to better distinguish between the internal state
variables and the user configurable settings. For the purposes of this
demo, the difference is largely superficial, but in a 'real world'
scenario where object orientation is an objective, this priv structure
would be opaque and inaccessible to the end user (i.e., declared in
particle-engine.c).