particle-initial: Use initial speed and direction, not vector
This decouples the particle's direction and speed by allowing both to be
specified separately. The direction should be applied in the form of a
unit vector which is then multiplied to the speed scalar to obtain the
initial velocity of the particle.

This has the added advantage that the amount of variance can better be
controlled, for example - it is possible to have a uniform particle
source direction with highly fuzzy speeds, or vice versa.