Add a particle-engine abstraction
This creates a generic particle engine type, and abstracts the details
of creating a painting a set of particles away form the logic involved
in creating and tracking their position. The idea here is that this will
then be used to support other types of particle systems, for example
particle swarms.